+++ lib/cog_api/http/groups.ex
@@ -1,0 +1,15 @@
+defmodule CogApi.HTTP.Groups do
+  alias CogApi.HTTP.Base
+
+  alias CogApi.Endpoint
+  alias CogApi.Resources.Group
+
+  def index(%Endpoint{}=endpoint) do
+    Base.get(endpoint, "groups") |> Base.format_response("groups", [%Group{}])
+  end
+
+  def create(%Endpoint{}=endpoint, params) do
+    Base.post(endpoint, "groups", %{group: params})
+    |> Base.format_response("group", %Group{})
+  end
+end