+++ lib/cog_api/resources/group.ex
@@ -1,0 +1,8 @@
+defmodule CogApi.Resources.Group do
+  @derive [Poison.Encoder]
+
+  defstruct [
+    :id,
+    :name,
+  ]
+end